+++ src/variables.ts
@@ -1,0 +1,4 @@
+export const initialModalSize = {
+  width: 400,
+  height: 250,
+};
